Add self-registration of stunt factories via assembly attributes
In order to support a pluggable mechanism for stunt factories (i.e. both C# source generator-based when supported, dynamic proxies elsewhere), we'll need a mechanism by which the DynamicProxy nuget will register itself to be the one to use, when built-time proxies aren't available.

The StuntFactoryAttribute provides this registration, and we make it user-friendly by checking in the package targets whether there are mulitple factories, which would not be a supported scenario.