Rename Acceptance>Integration tests add internals tests
The Roslyn internals access test is more integration in nature, so move to this project too, since it's typically going to take longer to run too since it needs to invoke nuget.org API and restore many packages for each run.

Also made that access test independent of the MSBuild APIs, using just `dotnet msbuild` instead, which simplifies the code a lot.