Ensure dependent assembly resolution can happen from the right dir
We don't want to unconditionally add all our generator dependencies as analyzers, since that might interfere with built-in ones that are already loaded (i.e. CodeAnalysis.Workspaces). So we instead pass along a property from targets with the analyzer directory containing the dependencies so we can resolve from there as needed.

NOTE: we cannot use the analyzer's assembly Location because that's coming from a temp directory created by Roslyn for the purposes of analyzer resolution and wouldn't allow us to track back to the net472/ne5.0 folders as needed.