Fix nullable warnings in generated code
Since the code we generate will depend in part of the scaffolding performed by the built-in codefixers, which do take into account the nullable attributes, we need to account for it to have nullable annotations. These annotations can only be emitted in generated source if it contains an explicit #nullable directive, so we add one and set it to always enabled since we can't know if the containing project will have that or not (or rather, it's easier this way ;)).

The code has already been carefully tested for nullability scenarios around the behavior pipeline, so we subsequently disable all nullable-related warnings since otherwise we'd have to generate different code for both cases (nullable enabled and disabled), and that doesn't really give us any benefits at this point.

We may revisit this decision in the future if it turns out we are throwning NREs from our generated stunts.