Simplify and just add CompilerGenerated to the classes
We were adding the attribute(s) to each generated member because we expected stunts to be generated by codefixers in the IDE, where users would regularly make partial classes of them. This is likely a super-corner scenario that we can introduce later if the need arises. For now it's better to keep things simple.