Fix tests reliability issues
The various ways we were constructing AdHocWorkspace in different tests
was causing random failures that were fixed upon re-runs, but this was
very annoying. After more investigation and unification into a single
way of retrieving the HostServices for the workspace, tests all pass
in a single run.

Restoring the GH workflow to test in that environment using just
dotnet test now.