Build debug config on all builds, except for release
Also attempt to fix code coverage with all that's in the web,
but it still seems to be reporting empty coverage :(. Even
locally, I can't get it to report anything.

Turn on the local file mapping, the source link support, and
the MSBuild collector (not that any of that made any difference
but supposedly it needs to be done?).