Allow optimizing behaviors collection change notifications
In Moq, for example, it will be common for the initial creation to set up a bunch of behaviors. The constant notifications one by one of each behavior addition would have a negative performance impact.

We implement the ISupportInitialize interface in a custom collection we use instead for the Behaviors observable collection.

Since both scenarios (listening for change notifications and changing in batches) are quite advanced scenarios, we also simplify the API surface for the pipeline/stunt Behaviors by making the collection a simple IList<IStuntBehavior>. This will be most natural and still preserves the described behavior, and the documentation remarks make now a note instead of those two advanced scenarios, which would require testing the instance for INotifyCollectionChanged/ISupportNotification interface implementations. Consumers of that behavior should check and fail gracefully if the collection does not implement either, although any stunt using the built-in codegen and our pipeline is guaranteed to implement both.