Bundle DynamicProxy in Stunts itself for convenience
At least initially, it doesn't make sense to force users to install the
Stunts.DynamicProxy when they are not on a C# 9 project does not seem like a great idea.

So we drop the standalone package, make it built-in via the Stunts.Package
project, and register the factory automatically whenever C# 9 isn't
supported.