Switch from attribute-based registration to partial static ctor
This makes things not only more discoverable and predictable, but also means a linker won't link away the dependency and it will perform better too since there is no reflection involved.

The default factory error message is now more explicit that there needs to be an assignment to StuntFactory.Default before usage.