Add strong-typed Get<T> for retrieving arguments
In order to make anonymous behaviors more concise but also provide better error messages when arguments aren't compatible with the requested types (instead of just throwing invalid cast exceptions), add a generic extension method that does various checks for value compatibility before casting and returning the value.

This also removes the need to explicitly add null check annotations and verifications, since you can request the type as nullable or not at the call site (i.e. `GetNullable<string>()` vs `Get<string>()` if you expect it to be not null).

The generated stunts use GetNullable always, but the resulting cast will fail if the argument type is a value type that is not nullable, resulting in a better error message than just the invalid cast we had before.