Add support for ref returns
C# 7+ supports ref returns and so does Stunts now :).

Callers can create a ref to a value by simply implicitly assigning a value to a variable of type Ref<T>:

    Ref<int> value = 42;

Then when creating the value return from a behavior, you can use that ref instead of the direct value (int 42 in this case), like:

    stunt.AddBehavior((invocation, next) => invocation.CreateValueReturn(value, invocation.Arguments.ToArray()));

If any code changes that ref value, it will be changed all the way back in your `value` variable above, like:

    ref int v = ref stunt.NumberAt(ref index, out var count);
    // this will cause the other variable to change too
    v = 42;

The `Ref<int>` has a `Value` property which is `ref return` too for all this to work.

The codegen was updated to account for this and will automatically "ref-box" a value if the method expects a ref returns but you only provided a direct value (i.e. an int). Obviously the by-ref semantics wouldn't be preserved in that case since callers would be modifying a copy of the value, not its original ref, but it makes this quite painless and straightforward to use if you're not actively monitoring the changes to the underlying value.

Fixes #23