Rename the project from Stunts to Avatar
I was inspired by the way Castle.DynamicProxy (see https://github.com/castleproject/Core/blob/master/docs/dynamicproxy-introduction.md#proxies) explains what proxies are and why they are useful using the metaphor of The Matrix.

It occurred to me that for the newer generations that might not be big fans of that awesome movie, some more familiar and recent might be better. It occurred to me that Avatar is equally good if not better as a metaphor: in The Matrix, everyone inside it are basically "proxies", which is not the way you'd typically use proxies and a library like this. Instead, in Avatar (the movie), only select few proxies are sent to an otherwise totally normal and real community of Na'vi (the rest of the components in your system that you're perhaps testing). Granted, the movie ending where Jake finally becomes a true Na'vi kinda ruins the metaphor, since a proxy will never be the real thing, but it's darn close :).

While explaining what I was doing to my son and wife in terms of the Avatar movie, I realized the library itself should just be named that. It's much easier to relate than Stunts (from Stunt Doubles > Test Doubles) and also removes the name from a test-only origin which isn't really accurate (since you can use proxies in production code too for a number of reasons).

So this renames things everywhere and adds a temporary new icon (the real one will come a bit later).