Add support for abstract and virtual class-based avatars
The support for calling the base implementation as the ultimate target for class-based pipelines was not implemented yet and all targets basically ended up in NotImplementedException. This adds the support by rewriting all members that are overrides of a virtual member (not abstract) and passes that as the ultimate target to the pipeline execution.