Use MethodInfo in IMethodInvocation instead of MethodBase
The reason we were using MethodBase was mostly because that's what MethodBase.GetCurrentMethod returns, and and because a constructor is also a MethodBase. We made assumptions (and safe downcasts too), however, that the method was mostly a MethodInfo.

This is because intercepting a constructor is not something we can really support, since that would require a behavior pipeline being configured *before* running the constructor. This would mean generating additional constructor(s) that can receive the pipeline as an argument (while generating the necessary constructor overloads).