Add base class constructor overriding and interception
The base class support we had previously didn't work if the base class had non-default constructors. This adds support for that by leveraging the same refactoring provide the IDE uses to generate all reachable constructors from the current class. This required extending the source codegen so that it can also run code actions, not just code fixes. The former come from the refactoring provider(s).

After that, the next logical step was to implement support for intercepting calls to the constructor, which requries a mechanism to get an initialized pipeline somehow *before* the object is constructed. For this purpose we introduce the IBehaviorPipelineFactory which follows a similar design like the IAvatarFactory. The static Default has been extended, however, to also allow a LocalDefault override, which uses an AsyncLocal<IBehaviorPipelineFactory> to allow per-flow overriding of the default pipeline factory. This should make unit testing much simpler in cases where that's needed.

Just like virtual members don't throw NotImplementedExceptions in their pipeline invocation because they provide an actual target (the base member call), ctor calls don't throw either, since they (by the very language design/rules) have always called the base constructor already by the time the pipeline is invoked. The generated code looks like the following:

```
public BaseType(string name): base(name) => pipeline.Execute(new MethodInvocation(this, MethodBase.GetCurrentMethod(), name), false);
```

The new `pipeline.Execute` overload receiving a boolean to not throw NIE assumes the method (in this case, ctor) is void (always the case for a ctor) and therefore creates a value return that isn't used. This means you can't use ctor interception to modify the arguments passed to the base class, for example. But you can record invocations, trigger callbacks, perform verifications and so on.

Note that unlike the built-in analyzers, which we can get from the composition as exports because they cannot have ctor dependencies, some built-in refactoring providers actually do have ctor dependencies that are not available in a source generator context, so instead we just get only the refactoring providers which have parameterles constructors, which happens to include the one we need to generate contructors from base class :).