Clean dependencies in Avatar.StaticProxy
Since the static proxy is intended to be bundled as into the
tools directory, we don't need it contributing public dependencies
on referencing packages. The MSBuild targets/properties that
determine whether it will be added as an analyzer or not are included
in the main package already.

This does mean that consumers of this as a standalone package
(for extensibility, like Moq.StaticProxy) will need to add their
own dependencies for all transitive deps, like Roslyn itself. But
that gives it also full control of which of those dependencies
propagate to the referencing projects too.