Improve design and usability of argument collection
This makes the argument collection a list of parameters with explicit Get/Set value methods by index and name, that properly validate that there is a backing parameter for each value.

We also make it easier to clone arguments collections for later modification by making the collection implementation public and adding object initializer overloads that allow for a more streamlined syntax like:

```csharp
var args = new ArgumentCollection(invocation.Arguments)
{
  { "message", "hello" },
};
```

(this case overrides just one of the parameters in the incoming invocation, for example, if there were more than one).

This sets the base for extension method-based extensibility of the marshaling of typed values into the collection, so that source-generated code will be able to Get/Set any typed value while enabling custom marshaling (such as for ref structs) without having to learn any new APIs.