⥱ Ensure compiler matches generator dependencies
Given that we pack and depend on some workspace APIs to perform some parts of the code generation, we need the compiler and our workspace dependencies to align. This is trivial to achieve by depending on the toolset that matches our Roslyn version.

This would not be required if we removed all dependencies on workspace APIs.

Fixes #52