◎ Simplify pipeline API by removing target, add callbase to IMethodInvocation
We previously had as part of the public pipeline API, the concept of an execution target, which would either be a `throw NotImplementedException` or the invocation of the base (for virtual) method.

This made the pipeline API a bit more complex than necessary, since knowing what the target is, is typically something that's hidden in the generated code for a proxy (or the dynamic proxy implementation).

The fact that the target was also not readily available in the method invocation itself, made it quite convoluted for a `CallBaseBehavior` (i.e. in Moq) to call it directly, and had instead to resort to skipping other behaviors (i.e. default and strict behaviors).

By moving the target to the `MethodInvocation` instead, we get to address two issues at once, both simplifying the `BehaviorPipeline.Execute` method as well as the scenario where one behavior wants to short-circuit a call by invoking the target directly. The `IMethodInvocation.SupportsCallBase` allows checking before invoking it (or filling in default, say).

Fixes #56.