Make MethodReturn public to easily allow modifying return values
Otherwise extenders would need to implement the (mostly trivial) class themselves, which doesn't add value and is just friction. All our other interfaces have public implementations anyway.

We make the class a record too, so it can easily be modified on the fly with `result with { Exception = ..., ReturnValue = ... }`, for example.

Fixes #61.