⌛ Run preview install in parallel, wait on build
To accelerate the workflow, we initiate the preview VS
install in parallel right from the start, saving significant
time since it's a lengthy operation. We poll for the status
of build and acceptance checks before continuing with the
preview tests.

Note we first wait on the build status, since when that isn't
run, the matrix of job names for acceptance won't be expanded
and we'd fail to get their status in turn. Once build completes
successfully, we can move to waiting on the acceptance tests,
since there's no point in running under VS preview if the
acceptance tests themselves don't pass either (in the stable
install of VS).