⧉ Avoid duplicate builds from Pack project (debug + release)
The version we were using of Traversal did not import the .NET SDK targets, and was missing the default configurations support that's now built-in. This caused the project (not having any default) to actually be built (packed) in Release mode, which caused all project references to be built in Release mode as well. If the solution was built with the Debug configuration, this meant everything was built twice :(.