Make arguments first-class and strong-typed
Previously we had code in quite a few places that would get the `ParameterInfo` from the argument collection, and then do something with the (untyped) object value subsequently. Since the collection was not typed, we were forcing boxing of every argument value, which in turn would prevent types that cannot be boxed in the future (such as Span<T>).

In addition, the validation of the argument value assignment according to its parameter info was quite loose and had to be accounted in more than one place too.

This commit adds a strong-typed `Argument<T>` record class (and its base un-typed class) which is now used consistently. It's also used to validate the type conversions so this is consistent regardless of which other APIs that end up creating argument collections are used.

In order to make it easy to create strong-typed argument collections, a bunch of generic factory method overloads were added to make this seamless:

* `ArgumentCollection.Create<...>(parameters, args)`
* `MethodInvocation.Create<...>`
* `IMethodInvocation.CreateResult<...>` and `CreateValueResult<...>` extension methods.

The first one is the actual one creating the typed arguments. The others are mostly pass-through to it, and are added as convenience to avoid having to invoke `Argument.Create<...>` explicitly, especially since the `ParameterInfo[]` to use in the collection is always the one in the method invocation itself.

Fixes #74.