⭾ Improve code formatting when emitting source files
Emitting source files by setting `EmitCompilerGeneratedFiles=true` in the project file is a very convenient way to explore and debug with full sources the output of source generators.

In order to make the experience even better, we could improve the formatting a bit, by properly adding missing newlines here and there, only when said flag is set (which would likely be while debugging/learning about a Avatar's internals).

Fixes #98.