Keep "manual" avatars updated from generated source
This makes them mostly an example of what generated avatars look like,
but otherwise functionally 100% equivalent to source-generated avatars.

Note that we don't pack avatars in ManualAvatars to avoid collisions.
Since the package is installed as source, the type names of generated
avatars would collide with the included "manual" avatars, so don't pack them.

We are including them in the source so they can be explored in Avatar'
source itself for learning/troubleshooting.