Rename callBase to implementation
In preparation for adding decorator support, it's becoming clearer that that callBase delegate is actually the implementation of a method, be it from a call to the base implementation in a virtual method or the invocation of a target object provided as the instance being decorated for an interface avatar.

In order to avoid having two distinct concepts for the same idea (i.e. in IMethodInvocation we have a CreateCallBaseReturn), rename things consistently so we only talk about the ultimate implementation of the method, so, HasImplementation and CreateInvokeReturn (to invoke that implementation).

When we introduce decorators, we will not need to add additional public API members in the pipeline for that.

Fixes #108.