Simplify package versioning and let dependabot do its work
Since dependabot does not know how to deal with the MSBuild we have (it breaks the Condition and nothing works from that point on), but it does know how to bump references in multiple files, just let it bump in more than one place and that's it.

Also, since both ThisAssembly and NuGetizer are fairly stable now, we don't need the additional 42.42.42 version conditioned to its presence in the local file system (for dogfooding). If needed, it can be manually tweaked locally and kept outside of the main repo.