Externalize default inline not implemented behavior
Instead of generating different invocation patterns depending on
whether there is an abstract/interface member implementation or
a virtual one, and throwing (inline) a NotImplementedException,
we instead enable avatars to have a target implementation (decorator-
style) that is invoked instead, which by default just happens to
throw NotImplementedException unless an alternative (now decorated)
implementation is provided.

This simplifies the generator as well as the target call stacks
which will always be consistent across abstract and interface-based
members, and also adds support for default interface implementations
since the default implementation would in that case not be generated
in the default target implementation.

Fixes #113.