Fix resource strings codegen for default resx behavior
The way we had switched to ThisAssemblyStrings, we sorta broke the sync of the built-in resx codegen, which now got out of sync and new strings wouldn't show up on the `Resources` class.

This fixes it by customizing the way the oss template changes it, via the Directory.targets extensibility point.