👀 Fail if no visible constructors available
If target type has no accessible constructor, report an error
diagnostic when code is generated. Currently, we need to do this
from the actual generator since we scaffold all default ctors
from there and it's an easy spot to query for them.

An analyzer would need to actually inspect the base type and
determine if generation would succeed at all. That would be a
nice perf. improvement if easy to implement, but right now it
does not seem worth it given the way we currently rely on code
fixes to scaffold the basic code.

Fixes #129.