Rename IAvatarProcessor to ISyntaxProcessor, cleanup context
In preparation for a generalized source generation/processing driver (not avatar-specific), rename the main interface for processors, since it will be used also for generating default (scaffold) implementation for non-avatar classes too.

Take the opportunity to remove the GeneratorAttribute from the processor context, since that's Avatar-specific too, and only used in the context of the syntax receiver to nominate candidate types for source generation. Make a missing generator type in the compilation a diagnostic error too (this was a TODO before).