Refactor existing syntax processor logic into new driver class
The new driver allows the code that drives the applying of syntax processors to be independent of other concerns, such as naming conventions and the like.

The new RoslynInternalScaffold re-implements the old behavior for scaffolding in a "standard" way within the new driver, so that no special-casing is needed at all.