Update to latest NuGet protocol package
The existing (now legacy) API was not returning the latest
prerelease package for some reason. The newer one returns
all entries as expected, so switch to that instead.

We also improve the output folder organization in the integration
test for internal access so that restoring one project does not
cause mismatched dependencies on another, since it seems otherwise
the project.assets.json ended up in the same location for all
projects, alongside the project file (which were previously all
at the same level).